+++ src/database/DataType.cpp
@@ -217,7 +217,7 @@ namespace db {
 			}
 
 			try {
-				size = std::stoi(strRepr.substr(i + 1, strRepr.size() - 1));
+				size = std::stoull(strRepr.substr(i + 1, strRepr.size() - 1));
 			} catch (const std::invalid_argument &e) {
 				throw UnknownDataTypeException("Size of data type \"" + strRepr
 											   + "\" could not be parsed: " + e.what());
